refactor(payment): 애플리케이션 계층 DTO 및 이벤트 금액 타입 Long → BigDecimal 변경
- Command/Result DTO 26종 금액 필드 변경
- SettlementCandidateCreatedEvent grossAmount 변경

+++ payment/src/main/java/com/example/payment/application/dto/CardPaymentConfirmCommand.java
@@ -1,11 +1,12 @@
 package com.example.payment.application.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CardPaymentConfirmCommand(
         UUID buyerId,
         UUID orderId,
         String paymentKey,
-        Long amount
+        BigDecimal amount
 ) {
 }
