+++ payment/src/main/java/com/example/payment/application/dto/EscrowReleaseResult.java
@@ -1,12 +1,13 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.EscrowStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record EscrowReleaseResult(
         UUID orderId,
-        Long releasedAmount,
+        BigDecimal releasedAmount,
         EscrowStatus escrowStatus,
         LocalDateTime releasedAt
 ) {
