+++ payment/src/main/java/com/example/payment/application/dto/EscrowTransactionItemResult.java
@@ -1,6 +1,7 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.EscrowTransactionType;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -12,9 +13,9 @@ public record EscrowTransactionItemResult(
         UUID sellerMemberId,
         UUID buyerMemberId,
         EscrowTransactionType transactionType,
-        Long amount,
-        Long beforeAmount,
-        Long afterAmount,
+        BigDecimal amount,
+        BigDecimal beforeAmount,
+        BigDecimal afterAmount,
         UUID referenceId,
         String referenceType,
         String description,
@@ -22,4 +23,3 @@ public record EscrowTransactionItemResult(
         LocalDateTime createdAt
 ) {
 }
-
