+++ payment/src/main/java/com/example/payment/application/dto/WithdrawResult.java
@@ -1,17 +1,18 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.WithdrawStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public record WithdrawResult(
         UUID withdrawRequestId,
-        Long amount,
-        Long fee,
-        Long actualAmount,
+        BigDecimal amount,
+        BigDecimal fee,
+        BigDecimal actualAmount,
         String maskedBankAccount,
         WithdrawStatus status,
-        Long walletBalance,
+        BigDecimal walletBalance,
         LocalDateTime requestedAt,
         LocalDateTime processedAt
 ) {
